pyproject.toml
[tool.poetry]
name = "mosques-accounting-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
